refactor(spdx-reporter): Move extension functions to a separate file
